requirements.txt
Django==3.1.5
django-crontab==0.7.1
djangorestframework==3.12.4
google-api-python-client==2.5.0

